+++ superuser_plugin_unix/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+* Uses binary search to determine current user is in admin user group.
+* Implement current user associated groups.
+
 ## 2.0.0-m.2
 
 * Expand error code length to two 16-bits, which lower 16-bit denote value of `errno` and upper 16-bit uses for category.